requirements.txt
anyio==3.6.2
appnope==0.1.3
appscript==1.2.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.1.0
async-generator==1.10
attrs==22.1.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
bs4==0.0.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coloredlogs==15.0.1
contourpy==1.0.6
cssselect==1.2.0
cycler==0.11.0
ddddocr==1.4.7
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.2.1
docopt==0.6.2
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.0.0rc9
executing==1.2.0
fake-useragent==0.1.14
fastjsonschema==2.16.2
flatbuffers==22.9.24
fonttools==4.38.0
h11==0.14.0
humanfriendly==10.0
idna==3.4
importlib-metadata==5.0.0
ipykernel==6.17.1
ipython==8.6.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.17.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-server==1.23.1
jupyter_client==7.4.5
jupyter_core==5.0.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
mistune==2.0.4
mpmath==1.2.1
nbclassic==0.4.8
nbclient==0.7.0
nbconvert==7.2.4
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.4
onnxruntime==1.13.1
opencv-python-headless==4.6.0.66
openpyxl==3.0.10
outcome==1.2.0
packaging==21.3
pandas==1.5.1
pandocfilters==1.5.0
parsel==1.7.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pipreqs==0.4.11
platformdirs==2.5.3
prettytable==3.5.0
prometheus-client==0.15.0
prompt-toolkit==3.0.32
protobuf==4.21.8
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pyecharts==1.9.1
pyee==8.1.0
Pygments==2.13.0
pymongo==4.3.3
PyMySQL==1.0.2
pyparsing==3.0.9
pyrsistent==0.19.2
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.6
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
requests==2.28.1
selenium==4.5.0
self==2020.12.3
Send2Trash==1.8.0
simplejson==3.17.6
six==1.16.0
sklearn==0.0.post1
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
stack-data==0.6.0
sympy==1.11.1
terminado==0.17.0
tinycss2==1.2.1
tornado==6.2
traitlets==5.5.0
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
w3lib==2.0.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
widgetsnbextension==4.0.3
wsproto==1.2.0
xlwings==0.28.5
yarg==0.1.9
zipp==3.10.0
